Source Code/Ingestion Pipeline/producer/requirements.txt
kafka-python==2.0.2
faker==23.2.1
python-dotenv==1.0.0
